Move Logic to logic package
